+++ shapedrawing.h
@@ -7,7 +7,8 @@ void sd_pathPoint(cairo_t *cr, gdouble x, gdouble y);
 void sd_pathLine(cairo_t *cr, gdouble x1, gdouble y1, gdouble x2, gdouble y2);
 
 void sd_pathArrow(cairo_t *cr, gdouble shapeXRef, gdouble shapeYRef,
-        gdouble shapeXEnd, gdouble shapeYEnd, gdouble thickness, gint angle);
+        gdouble shapeXEnd, gdouble shapeYEnd, gdouble thickness,
+        gdouble angle);
 
 void sd_pathTriangle(cairo_t *cr, gdouble xBeg, gdouble yBeg,
         gdouble xEnd, gdouble yEnd, gdouble angle, gdouble round);
